PR Comments Improvements
