+++ src/SFA.DAS.Tools.Support.UnitTests/ApprovalsControllerTests.cs
@@ -3,9 +3,7 @@ using AutoFixture.Xunit2;
 using FluentAssertions;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Options;
 using Moq;
-using NLog.Targets;
 using SFA.DAS.Tools.Support.Core.Models;
 using SFA.DAS.Tools.Support.Infrastructure.Services;
 using SFA.DAS.Tools.Support.UnitTests.AutoFixture;
@@ -15,7 +13,6 @@ using SFA.DAS.Tools.Support.Web.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
@@ -39,84 +36,6 @@ namespace SFA.DAS.Tools.Support.UnitTests
         }
 
         [Theory, AutoMoqData]
-        public async Task SearchApprenticeships_POST_EmptyModel_ReturnsError(ApprovalsController sut)
-        {
-            //Given
-            var model = new SearchApprenticeshipsViewModel();
-
-            //When
-            var result = await sut.SearchApprenticeships(model);
-
-            //Then
-            result.Should().BeOfType<JsonResult>().Which
-                .Value.Should().BeEquivalentTo(new
-                {
-                    ErrorTitle = "Invalid Search",
-                    ErrorMessage = "At least one parameter must be populated"
-                });
-        }
-
-        [Theory, AutoMoqData]
-        public async Task SearchApprenticeships_POST_ApiCallErrors_ReturnsError([Frozen] Mock<IEmployerCommitmentsService> api, SearchApprenticeshipsViewModel model, ApprovalsController sut)
-        {
-            //Given
-            var apiResult = new SearchApprenticeshipsResult
-            {
-                ErrorMessage = "Api Error Message"
-            };
-
-            api.Setup(s => s.SearchApprenticeships(
-                It.Is<SearchApprenticeshipsRequest>(s => s.CourseName == model.CourseName && s.EmployerName == model.EmployerName
-                && s.ProviderName == model.ProviderName && s.StartDate == model.StartDate && s.EndDate == model.EndDate && s.SearchTerm == model.ApprenticeName
-                && s.ApprenticeshipStatus == model.SelectedStatus),
-                It.IsAny<CancellationToken>()))
-             .Returns(Task.FromResult(apiResult));
-
-            //When
-            var result = await sut.SearchApprenticeships(model);
-
-            //Then
-            result.Should().BeOfType<JsonResult>().Which
-                .Value.Should().BeEquivalentTo(new
-                {
-                    ErrorTitle = "Call to Commitments Api Failed",
-                    ErrorMessage = apiResult.ErrorMessage
-                });
-        }
-
-        [Theory, AutoMoqData]
-        public async Task SearchApprenticeships_POST_ApiCallSucceeds_ReturnsResult([Frozen] Mock<IEmployerCommitmentsService> api, SearchApprenticeshipsViewModel model, SearchApprenticeshipsResult apiResult, ApprovalsController sut)
-        {
-            //Given
-            apiResult.ErrorMessage = "";
-            api.Setup(s => s.SearchApprenticeships(
-                It.Is<SearchApprenticeshipsRequest>(s => s.CourseName == model.CourseName && s.EmployerName == model.EmployerName
-                && s.ProviderName == model.ProviderName && s.StartDate == model.StartDate && s.EndDate == model.EndDate && s.SearchTerm == model.ApprenticeName
-                && s.ApprenticeshipStatus == model.SelectedStatus),
-                It.IsAny<CancellationToken>()))
-             .Returns(Task.FromResult(apiResult));
-
-            //When
-            var result = await sut.SearchApprenticeships(model);
-
-            //Then
-            result.Should().BeOfType<JsonResult>().Which
-                .Value.Should().BeEquivalentTo(apiResult.Apprenticeships.Select(s => new
-                {
-                    s.Id,
-                    s.FirstName,
-                    s.LastName,
-                    s.EmployerName,
-                    s.ProviderName,
-                    s.CourseName,
-                    s.StartDate,
-                    s.EndDate,
-                    Status = s.ApprenticeshipStatus.ToString(),
-                    PaymentStatus = s.PaymentStatus.ToString()
-                }));
-        }
-
-        [Theory, AutoMoqData]
         public async Task StopApprenticeship_POST_NoRowsSelected_RedirectsToSearch(ApprovalsController sut, ApprenticeshipSearchResultsViewModel model)
         {
             //Given
@@ -127,7 +46,7 @@ namespace SFA.DAS.Tools.Support.UnitTests
 
             //Then
             var action = result.Should().BeOfType<RedirectToActionResult>().Which;
-            action.ActionName.Should().Be(ApprovalsRouteNames.SearchApprenticeships);
+            action.ActionName.Should().Be(RouteNames.Approval_SearchApprenticeships);
             action.RouteValues.Values.Should().BeEquivalentTo(new[]
             {
                     model.ApprenticeName,
@@ -253,7 +172,7 @@ namespace SFA.DAS.Tools.Support.UnitTests
 
             //Then
             var action = result.Should().BeOfType<RedirectToActionResult>().Which;
-            action.ActionName.Should().Be(ApprovalsRouteNames.SearchApprenticeships);
+            action.ActionName.Should().Be(RouteNames.Approval_SearchApprenticeships);
             action.RouteValues.Values.Should().BeEquivalentTo(new[]
             {
                     model.SearchParams.ApprenticeName,
