--- src/SFA.DAS.Tools.Support.Web/Configuration/ApprovalsRouteNames.cs
@@ -1,11 +1,0 @@
-﻿namespace SFA.DAS.Tools.Support.Web.Configuration
-{
-    public class ApprovalsRouteNames
-    {
-        public const string StopApprenticeship = "StopApprenticeship";
-        public const string CancelStopApprenticeship = "CancelStopApprenticeship";
-        public const string StopApprenticeshipConfirmation = "StopApprenticeshipConfirmation";
-        public const string SearchApprenticeships = "SearchApprenticeships";
-        
-    }
-}
