+++ src/SFA.DAS.Tools.Support.Web/Configuration/RouteNames.cs
@@ -1,0 +1,13 @@
+﻿namespace SFA.DAS.Tools.Support.Web.Configuration
+{
+    public class RouteNames
+    {
+        public const string Approval_StopApprenticeship = "StopApprenticeship";
+        public const string Approval_CancelStopApprenticeship = "CancelStopApprenticeship";
+        public const string Approval_StopApprenticeshipConfirmation = "StopApprenticeshipConfirmation";
+        public const string Approval_SearchApprenticeships = "SearchApprenticeships";
+
+        public const string CommitmentsData_SearchApprenticeships = "Search";
+        
+    }
+}
